Split cmux events CLI helpers
